+++ core/client/assets/sass/helpers/variables.scss
@@ -45,6 +45,8 @@ $at2x: 2 device-pixel-ratio;
 
 $dropdown_triangle: 8px;
 
+$popover_triangle: 14px;
+$popover_triangle_shallow_multiplier: 0.8;
