+++ core/client/assets/sass/screen.scss
@@ -43,6 +43,7 @@
 @import "components/dropdowns";
 @import "components/pagination";
 @import "components/badges";
+@import "components/popovers";
 
 
 //
