+++ core/client/components/gh-popover.js
@@ -1,0 +1,7 @@
+import GhostDropdown from 'ghost/components/gh-dropdown';
+
+var GhostPopover = GhostDropdown.extend({
+    classNames: 'ghost-popover'
+});
+
+export default GhostPopover;
