+++ core/client/initializers/dropdown.js
@@ -1,17 +1,4 @@
-import BodyEventListener from 'ghost/mixins/body-event-listener';
-
-var DropdownService = Ember.Object.extend(Ember.Evented, BodyEventListener, {
-    bodyClick: function (event) {
-        /*jshint unused:false */
-        this.closeDropdowns();
-    },
-    closeDropdowns: function () {
-        this.trigger('close');
-    },
-    toggleDropdown: function (dropdownName, dropdownButton) {
-        this.trigger('toggle', {target: dropdownName, button: dropdownButton});
-    }
-});
+import DropdownService from 'ghost/utils/dropdown-service';
 
 var dropdownInitializer = {
     name: 'dropdown',
@@ -19,10 +6,16 @@ var dropdownInitializer = {
     initialize: function (container, application) {
         application.register('dropdown:service', DropdownService);
 
+        // Inject dropdowns
         application.inject('component:gh-dropdown', 'dropdown', 'dropdown:service');
         application.inject('component:gh-dropdown-button', 'dropdown', 'dropdown:service');
         application.inject('controller:modals.delete-post', 'dropdown', 'dropdown:service');
         application.inject('controller:modals.transfer-owner', 'dropdown', 'dropdown:service');
+        application.inject('route:application', 'dropdown', 'dropdown:service');
+
+        // Inject popovers
+        application.inject('component:gh-popover', 'dropdown', 'dropdown:service');
+        application.inject('component:gh-popover-button', 'dropdown', 'dropdown:service');
         application.inject('route:application', 'dropdown', 'dropdown:service');
     }
 };
