+++ core/client/templates/debug.hbs
@@ -6,7 +6,7 @@
 <div class="page-content">
 
     <div class="settings-content">
-        
+
         <header class="settings-view-header">
             <h2 class="page-title">Ugly Debug Tools</h2>
             <div class="js-settings-header-inner settings-header-inner">
@@ -15,6 +15,7 @@
         </header>
 
         <section class="content settings-debug">
+
             <form id="settings-export">
                     <fieldset>
                         <div class="form-group">
